+++ drake/systems/framework/test/state_subvector_test.cc
@@ -13,8 +13,8 @@ namespace drake {
 namespace systems {
 namespace {
 
-const size_t kLength = 4;
-const size_t kSubVectorLength = 2;
+const int64_t kLength = 4;
+const int64_t kSubVectorLength = 2;
 
 class StateSubvectorTest : public ::testing::Test {
  protected:
